src/resources/whisperServer/src/requirements.txt
av==11.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.2.1
faster-whisper==1.0.1
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.3.1
grpcio==1.62.2
grpcio-tools==1.62.2
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.24.4
onnxruntime==1.17.3
packaging==24.0
protobuf==4.25.3
PyYAML==6.0.1
requests==2.31.0
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
webrtcvad==2.0.10
